Refactor bezier lib file into a separate folder
