+++ libraries/bezier-rs/src/bezier/structs.rs
